deleted stuff - should further confirm full functionality
