--- index4.js
@@ -1,49 +1,0 @@
-const globalMethods = require("./globalMethods");
-var userdb = require("./db");
-require('dotenv').config();
-const fs = require('node:fs');
-const path = require('node:path');
-const axios = require('axios');
-var Datastore = require(`nedb`);
-// TODO: break in case of emergency (uncomment below line for possible db fix)
-// var db = new Datastore({filename: `users.db`});
-
-// importing the items we need from discord.js package
-const { Client, Collection, Events, GatewayIntentBits } = require('discord.js');
-//configuring events the bot can recieve
-const client = new Client({ intents: [GatewayIntentBits.Guilds, GatewayIntentBits.GuildMessages, GatewayIntentBits.MessageContent] });
-
-client.commands = new Collection();
-
-const commandsPath = path.join(__dirname, 'commands');
-const commandFiles = fs.readdirSync(commandsPath).filter(file => file.endsWith('.js'));
-
-for (const file of commandFiles) {
-    const filePath = path.join(commandsPath, file);
-    const command = require(filePath);
-    // Set a new item in the Collection with the key as the command name and the value as the exported module
-    if ('data' in command && 'execute' in command) {
-        client.commands.set(command.data.name, command);
-    } else {
-        console.log(`WARNING The command at ${filePath} is missing a required "data" or "execute" property.`)
-    }
-}
-
-// client.on('ready', () => {
-//     console.log('bot is ready');
-// })
-
-const eventsPath = path.join(__dirname, 'events');
-const eventFiles = fs.readdirSync(eventsPath).filter(file => file.endsWith('.js'));
-
-for (const file of eventFiles) {
-	const filePath = path.join(eventsPath, file);
-	const event = require(filePath);
-	if (event.once) {
-		client.once(event.name, (...args) => event.execute(...args));
-	} else {
-		client.on(event.name, (...args) => event.execute(...args));
-	}
-}
-
-client.login(process.env.DISCORD_BOT_ID);
