--- nedb_struggles.txt
@@ -1,16 +1,0 @@
-ENOENT error - deifnitely caused by multiple files opening the users.db at the same time
- 
-We don’t need a new API to have singletons, just abstract it with a local module of your own, e.g.
-db.js:
-const Datastore = require('nedb');
-const db = {
-  tableName: new Datastore({ filename: '...' })
-};
-module.exports = db;
-    
-    -Great thinking so far - need a way to point to parent folder on import in the command files
-
-I met the same problem, but i found that error appears when creating multiple instances of datastore with same filename. 
-You have to instantialize datastore only once and keep reference across whole app. 
-May be developers should provide singleton/factory 
-(for example var db = require('nedb').open({filename: 'path/to/file'}) ) instead of creating instances
